pyproject.toml
[project]
name = "cms-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn==0.30.0",
    "python-multipart==0.0.7",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt,argon2]==1.7.4",
    "sqlalchemy==2.0.31", # keep only the latest one you listed
    "pymysql==1.1.0",
    "alembic==1.11.1",
    "pydantic==2.8.2",
    "pydantic-settings==2.3.4",
    "pydantic-core==2.20.1",
    "slowapi==0.1.9",
    "loguru==0.7.3",
    "faker==37.8.0",
    "bcrypt==4.0.1",
    "python-dotenv>=1.1.1",
]
